Split wave composer component responsibilities
